sdimage: fix printf warnings about expected argument types
Depending on compiler used, two warnings are presented for the printf
statement, e.g.

sdimage.c:308:28: warning: format '%ld' expects argument of type 'long int',
 but argument 2 has type 'off_t {aka long long int}' [-Wformat=]

However, off_t depends on compiler and platform and there is no
portable way (like %zd for size_t type) to print off_t type.

So cast it to long long and use %lld instead to fix the mess.

+++ sdimage.c
@@ -305,7 +305,8 @@ int main(int argc, char *argv[])
 	}
 
 	if (verbose > 1) {
-		printf("Firmware size: %ld bytes, %ld sectors\n", fw_stat.st_size, SECTOR_COUNT(fw_stat.st_size));
+		printf("Firmware size: %lld bytes, %lld sectors\n",
+		       (long long int)fw_stat.st_size, (long long int)SECTOR_COUNT(fw_stat.st_size));
 	}
 
 	/* open target device and read MBR with partition table */
